requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
langchain==0.3.20
langchain-community==0.3.19
llama-cpp-python==0.3.8
requests==2.32.3
beautifulsoup4==4.13.3
sqlalchemy==2.0.39
pydantic==2.10.6
asyncpg==0.30.0
pydantic-settings==2.8.1
python-dotenv==1.0.1